refactor to use docker compose sdk instead of cli
